Move socket and pid files in SUSE defaults (bsc#1089770)
MariaDB in SUSE expect some files (like the socket) in /run/mysql
directory.  Some SUSE components are hardcoded under those
expectations (like YaST), and fail if this assumption is wrong.

This patch move the scoket and the pid files under /run/mysql, and
use more Jinja templates in places where it is expected.

Change-Id: Ic09d294e6a7d3b2fd65367369c5ef45d0b3010d9