Quick check to reduce Percona confusion (bsc#1091492)
There has been a lot of confusion about how to bootstrap the
MariaDB cluster in Cloud 8.  Unfortunately we didn't rename the
playbooks, so many users try to run percona-bootstrap.yml, which
doesn't work for Cloud 8 (leftover from HOS).
This is just one small step in the right direction to avoid
complications from running the wrong script by failing on a SUSE
install.

Change-Id: I583677084ceb291c39c8bf9750278a002c3bb55f