use galera-bootstrap for galera cluster (bsc#1091869)
For Galera cluster, we must use the Galera bootstrap tasks
instead of the old Percona logic.

This patch also remove the old Percona logic as it is no longer
supported. For example, galera-boot will handle the starts of all
the nodes. Therefore, we don't need to include start.yml anymore.

Change-Id: Ic228896a6507dcd0123b3ddbb68cacd1d696bfb2