Use galera-upgrade.yml for update/upgrade workflow (SOC-10521)
For the update/upgrade workflow the galera-upgrade.yml playbook will
be running after we have use zypper to patch/update/dist-upgrade the
packages on a node, so all we really need to do is ensure that any
additional packages, not implicitly installed via dependencies, get
installed, and then ensure that the database configuration files are
updated and that the MariaDB/Galera instances are (re)started to use
the updated packages and config settings.

In the case where we are running for all nodes in the FND-MDB host
group, as is the case during the upgrade workflow, or when running
the update workflow for all nodes using the skip_single_host_checks
setting to disable single hosts checks, the existing task list in
roles/FND-MDB/tasks/galera_bootstrap.yml handles most of this for
us, though we do need to revise it somewhat to avoid issues seen
with using the entirety of the hostvars in larger plays spanning
multiple services leading to ansible undefined fact errors because
dependent service roles are not included in the play context.

In the case where we are running for a single node, using --limit,
we need to run the configure and start task lists to ensure that any
new configuration changes are applied and that the service instances
are restarted.

However we do not support running this playbook for a more than one
node, but less than all nodes in the FND-MDB host group, and will
fail with an appropriate error if this scenario is detected.

When sorting versions we can use the --version-sort option rather than
the numeric sort which may not always do the right thing.

Also cleanup a couple of comment typos and change some user visible
'percona' references to 'galera'.

Change-Id: I527d683ec5bc27903dfae2a0d8205d0de7b5aae5