account for pre-update nodes (SOC-11037)
To support the scenario where not all the mysql nodes are updated and
we need to recover the cluster. In this case, we need to account for
the old mysql config file for the nodes that hasn't been updated yet.

There are accounting for the following scenarios:

1) New install. In this scenario, the new mysql conf file location
is used as the old mysql conf file should not have been there.
2) No update. In this scenario, the node hasn't been updated yet and
we need to recover the cluster. In this case, the new conf file location
should not have existed.
3) Updated. In this scenario, the new mysql conf file location should be
there and it should take precedence over the old location.
4) Corrupted. galera_bootstrap.yml play was executed and the node hasn't been
updated yet. In this case, we ended with both the new and old config files, and
the old config file is overriding the new config file.

Change-Id: I85ba8adfe880dcc80704d2b92ce545841ee88488