Switch to using override file in my.cnf.d (SOC-11043)
Instead of trying to edit the existing conf file when we need to
override the value of the wsrep_cluster_address setting during the
bootstrap process, we can create a file that sorts lexically after
all the other conf files, e.g. starting with 'zzz-', with the
desired override setting, and then delete it once the bootstrap
process has completed.

This eliminates the need to orchestrate when the standard conf files
are generated, and works for both the old and new on-disk config file
layouts with out the need for any special handling.

Change-Id: Idf2e8b69861f188c73f80afa8e8619c0f7cc781a