Improve boostrap error handling (SOC-11207)
The underlying error message for the
find_galera_mariadb_bootstrap_host Ansible module did not correctly get
conveyed to stdout. This patch fixed that behavior.

Also, in the situation where the mysql process is running but the seqno
is not -1 as expected due to possible SST problems, we need to advice
customer to check the logs on that node and may need to manually restart
the mysql process in order to resolve the problem.

Change-Id: I8029a4a3fe11b396acbd68c85bd0ee747cff116a